🚚 rename: useReservationDetails 파일 이동 및 페치 로직 분리 #93
